Make USE_AVX flag default value OFF (#3983)
With the emergence of new ISA (like ARM and RISCV) retaining USE_AVX ON default makes the default build instructions fail. Fundamentally DGL does not require the use of AVX for functional working. AVX is mainly needed when to enable optimization. So proposal is to default turn it off and then later during build instructions, folks with AVX capabilities can enable with 
`cmake .. -DUSE_AVX=ON`

Co-authored-by: Zihao Ye <expye@outlook.com>